requirements.txt
#Comandos para iniciar la app
#python manage.py tailwind start
#python manage.py runserver
django
pillow==10.4.0
mysqlclient==2.2.4
django-environ==0.11.2
djlint==1.35.2
django-tailwind[reload]
django-tinymce